Chnage all int types to long

+++ src/SalesManager.java
@@ -15,9 +15,9 @@ public class SalesManager {
         return max;
     }
 
-    public int min(){
-        int min = Integer.MAX_VALUE;
-        for(int sale : sales){
+    public long min(){
+        long min = Long.MAX_VALUE;
+        for(long sale : sales){
             if (sale < min){
                 min = sale;
             }
@@ -25,14 +25,14 @@ public class SalesManager {
         return min;
     }
 
-    public int getMiddleValueIgnoreHighestAndLowest(){
-        int min = min();
-        int max = max();
+    public long getMiddleValueIgnoreHighestAndLowest(){
+        long min = min();
+        long max = max();
 
-        int sum = 0;
-        int amountOfSales = 0;
+        long sum = 0;
+        long amountOfSales = 0;
 
-        for(int sale : sales){
+        for(long sale : sales){
             if(sale != min && sale != max){
                 sum += sale;
                 amountOfSales++;
